Cache static assets and handle download failures
