+++ app.py
@@ -1,40 +1,15 @@
-import streamlit as st, pandas as pd, numpy as np, plotly.express as px, unicodedata, re, json, requests
-from pathlib import Path
+import streamlit as st, pandas as pd, numpy as np, plotly.express as px, unicodedata, re
+from utils import get_logo_bytes, get_geojson
 
 st.set_page_config(page_title="Interventions Enedis", layout="wide", initial_sidebar_state="expanded")
 
-root = Path(__file__).parent
-logo = root / "enedis_logo.png"
-geo = root / "arrondissements.geojson"
-
-
-def _download(url: str, dest: Path, timeout: int = 15) -> None:
-    """Download a file to dest, handling network errors gracefully."""
-    try:
-        r = requests.get(url, timeout=timeout)
-        r.raise_for_status()
-    except Exception as e:
-        st.warning(f"Impossible de télécharger {url} : {e}")
-        return
-    dest.write_bytes(r.content)
-
-
-if not logo.exists():
-    _download(
-        "https://upload.wikimedia.org/wikipedia/fr/7/7d/Enedis_logo.svg",
-        logo,
-        timeout=15,
-    )
-if not geo.exists():
-    _download(
-        "https://opendata.paris.fr/explore/dataset/arrondissements/download/?format=geojson",
-        geo,
-        timeout=30,
-    )
-
 enedis_cols = ["#2C75FF", "#75C700", "#4A9BFF", "#A0D87C", "#0072F0", "#47B361", "#6EABFF", "#9EE08E"]
 
-st.image(str(logo), width=220)
+logo_bytes = get_logo_bytes()
+if logo_bytes:
+    st.image(logo_bytes, width=220)
+else:
+    st.warning("Logo manquant")
 
 def _n(x):
     return ''.join(c for c in unicodedata.normalize('NFKD', str(x)) if not unicodedata.combining(c)).lower().replace(' ', '').replace('_', '')
@@ -296,11 +271,11 @@ if {"Temps théorique", "Temps réalisé", "Prestation"}.issubset(flt.columns):
     f = px.bar(t, x="Prestation", y=["Temps théorique", "Temps réalisé"], color_discrete_sequence=enedis_cols[:2], barmode="group", title="Temps théorique vs réalisé par prestation")
     st.plotly_chart(f, use_container_width=True)
 
-if "Arr" in flt.columns and geo.exists():
+gj = get_geojson()
+if "Arr" in flt.columns and gj:
     arr = flt["Arr"].value_counts().rename_axis("Arr").reset_index(name="n")
     arr["Arr"] = arr["Arr"].astype(int)
     arr["pct"] = pct(arr["n"])
-    gj = json.loads(geo.read_text())
     f = px.choropleth(
     arr,
     geojson=gj,
